Changelog ui color and appearance (#3038)
* Changelog ui color and appearance

* enhance: Refactor changelog page with modular components (SearchBar, ChangeBadge, PackageToggle)

---------
